Add create new canvas UI
This is not a fully functional feature. The "creation" is simply a navigation to the MainView. There is no new canvas added to the canvases state
in CanvasSelectionView. This functionality is to be implemented together with the naming of the Canvas. This draws inspiration from Notability
where a canvas is only actually created once there is a CanvasElement added or the title of the canvas is changed, otherwise, a backward
navigation will result in the new canvas not being saved.