Refactor backlink component
Previous implementation had a lot of unnecesary "layering" and
complexity. It also had difficulties supporting encoding of
the models for persistence storage since the Node and Edge objects of
the Graph library will not be exposed to the main Dynavity application.
With that, I've created another BacklinkNode and BacklinkEdge class,
which will be exposed to the main Dynavity application and can be
encoded / decoded.